// ClientApp/src/Entity/Enums/MenuEnums.ts
export enum HeaderEnums {
    Posts = "posts",
    TypeOfNews = "typeOfNews",
    Advertisements = "advertisements",
    Accounts = "accounts",
    Roles = "roles",
    Crawler = "crawlers",
    Video = "video",
}

export enum MenuEnums {
    Add = "add",
    Send = "send",
    Hidden = "hidden",
    Upload = "upload",
    Position = "position",
    Download = "download",
    Delete = "delete",
    Vote = "vote",
    List = "list",
    Null = "Null",
    AddRole = "addrole",
    Folders = "folders"
}